Data.Promise

Description

Lazy promises

Documentation

data Lazy s a

runLazy :: (forall s. Promise s a -> Lazy s ()) -> a -> a

runLazy_ :: (forall s. Promise s a -> Lazy s ()) -> a

data Promise s a where

A lazy I-Var.

Constructors

Promise :: MVar a -> a -> Promise s a

promise :: a -> Lazy s (Promise s a)

Promise that by the end of the computation we'll provide a "real" answer, or we'll fall back and give you this answer

promise_ :: Lazy s (Promise s a)

Promise that by the end of the computation we'll provide a "real" answer, or you'll get an error.

(!=) :: Promise s a -> a -> Lazy s () infixl 0

Fulfill a promise.

demand :: Promise s a -> a

Demand the result of an I-Var

data BrokenPromise

Constructors

BrokenPromise